删除

+++ wfront/component/service.py
@@ -16,6 +16,9 @@ def w_component_service(root,ojson):
 		for component in components:
 			component_name = component['table']
 			module = component['module']
+			table = databases_dir[component_name]
+			crud = table.get('crud')
+			module = component['module']
 			os.makedirs(os.path.join(root,f'src/pages/{path}/{component_name.lower()}_{module}'),exist_ok=True)
 			initdir = os.path.join(root,f'src/pages/{path}/{component_name.lower()}_{module}/service.ts')
 			w = open(initdir,'w+')
@@ -26,37 +29,38 @@ def w_component_service(root,ojson):
 
 			w.write(f"""export async function query{component_name}List(params?: TableListParams) {{\n""")
 			w.write(f"""  return request('/api/{component_name.lower()}/list', {{\n""")
-			w.write(f"""    method: 'POST',\n""")
-			w.write(f"""    data: {{\n""")
-			w.write(f"""      ...params,\n""")
-			w.write(f"""    }},\n""")
+			w.write(f"""    params,\n""")
 			w.write(f"""  }});\n""")
+			w.write(f"""}}\n""")
+			w.write(f"""\n""")
+
+			if "post" in crud:
+				w.write(f"""export async function add{component_name}(params: TableListParams) {{\n""")
+				w.write(f"""  return request('/api/{component_name.lower()}', {{\n""")
+				w.write(f"""    method: 'POST',\n""")
+				w.write(f"""    data: {{\n""")
+				w.write(f"""      ...params,\n""")
+				w.write(f"""    }},\n""")
+				w.write(f"""  }});\n""")
+				w.write(f"""}}\n""")
+				w.write(f"""\n""")
+
+			if "delete" in crud:
+				w.write(f"""export async function remove{component_name}(params: {{ key: number[] }}) {{\n""")
+				w.write(f"""  return request('/api/{component_name.lower()}', {{\n""")
+				w.write(f"""    method: 'DELETE',\n""")
+				w.write(f"""    data: {{\n""")
+				w.write(f"""      ...params,\n""")
+				w.write(f"""      method: 'delete',\n""")
+				w.write(f"""    }},\n""")
+				w.write(f"""  }});\n""")
+				w.write(f"""}}\n""")
+				w.write(f"""\n""")
 
 
-			w.write(f"""}}\n""")
-			w.write(f"""\n""")
-			w.write(f"""export async function removeRule(params: {{ key: number[] }}) {{\n""")
-			w.write(f"""  return request('/api/rule', {{\n""")
-			w.write(f"""    method: 'POST',\n""")
-			w.write(f"""    data: {{\n""")
-			w.write(f"""      ...params,\n""")
-			w.write(f"""      method: 'delete',\n""")
-			w.write(f"""    }},\n""")
-			w.write(f"""  }});\n""")
-			w.write(f"""}}\n""")
-			w.write(f"""\n""")
-			w.write(f"""export async function addRule(params: TableListParams) {{\n""")
-			w.write(f"""  return request('/api/rule', {{\n""")
-			w.write(f"""    method: 'POST',\n""")
-			w.write(f"""    data: {{\n""")
-			w.write(f"""      ...params,\n""")
-			w.write(f"""      method: 'post',\n""")
-			w.write(f"""    }},\n""")
-			w.write(f"""  }});\n""")
-			w.write(f"""}}\n""")
-			w.write(f"""\n""")
-			w.write(f"""export async function updateRule(params: TableListParams) {{\n""")
-			w.write(f"""  return request('/api/rule', {{\n""")
+
+			w.write(f"""export async function update{component_name}(params: TableListParams) {{\n""")
+			w.write(f"""  return request('/api/{component_name.lower()}', {{\n""")
 			w.write(f"""    method: 'POST',\n""")
 			w.write(f"""    data: {{\n""")
 			w.write(f"""      ...params,\n""")
